+++ source/Flag.js
@@ -2,13 +2,12 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
 
-import getFlagCountryCode from './getFlagCountryCode'
-
 // Default country flag icon.
 // `<img/>` is wrapped in a `<div/>` to prevent SVGs from exploding in size in IE 11.
 // https://github.com/catamphetamine/react-phone-number-input/issues/111
 export default function FlagComponent({
 	country,
+	countryName,
 	flags,
 	flagUrl,
 	className
@@ -18,9 +17,9 @@ export default function FlagComponent({
 	}
 	return (
 		<img
-			alt={country}
+			alt={countryName}
 			className="PhoneInputCountryIconImg"
-			src={flagUrl.replace('{0}', getFlagCountryCode(country).toLowerCase())}/>
+			src={flagUrl.replace('{0}', country.toLowerCase())}/>
 	)
 }
 
@@ -29,9 +28,12 @@ FlagComponent.propTypes = {
 	// Two-letter country code ("ISO 3166-1 alpha-2").
 	country: PropTypes.string.isRequired,
 
+	// Will be HTML `title` attribute of the `<img/>`.
+	countryName: PropTypes.string.isRequired,
+
 	// Country flag icon components.
 	// By default flag icons are inserted as `<img/>`s
-	// with their `src` pointed to `flag-icon-css` github repo.
+	// with their `src` pointed to this library's github pages website.
 	// There might be cases (e.g. an offline application)
 	// where having a large (3 megabyte) `<svg/>` flags
 	// bundle is more appropriate.
@@ -39,6 +41,6 @@ FlagComponent.propTypes = {
 	flags: PropTypes.objectOf(PropTypes.elementType),
 
 	// A URL for a country flag icon.
-	// By default it points to `flag-icon-css` github pages website.
+	// By default it points to this library's github pages website.
 	flagUrl: PropTypes.string.isRequired
 }